refactoring

+++ app/src/main/java/com/jimbodev/jimla/projectg/Boid.java
@@ -15,7 +15,7 @@ class Boid extends Mobile implements Runner{
 
     private int health = 100;
 
-    Boid(float x, float y, float size, Node startNode, Node goalNode) {
+    Boid(float x, float y, int size, Node startNode, Node goalNode) {
         super(x, y, size, size, new Random().nextInt(10) + 2, 5, null);
 
         this.nextTargetNode = startNode;
